# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication & Registration"
                ],
                "summary": "User log in",
                "description": "Returns user data with token",
                "operationId": "login",
                "requestBody": {
                    "$ref": "#/components/requestBodies/LoginRequest"
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Authentication & Registration"
                ],
                "summary": "User registration",
                "operationId": "register",
                "requestBody": {
                    "$ref": "#/components/requestBodies/RegisterRequest"
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Authentication & Registration"
                ],
                "summary": "User log out",
                "description": "Method revokes the current access token.",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reset-password-request": {
            "post": {
                "tags": [
                    "Authentication & Registration"
                ],
                "summary": "User reset password request. Send email with reset link.",
                "operationId": "resetPasswordRequest",
                "requestBody": {
                    "$ref": "#/components/requestBodies/RequestResetPasswordRequest"
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            }
        },
        "/api/reset-password": {
            "post": {
                "tags": [
                    "Authentication & Registration"
                ],
                "summary": "User change password",
                "operationId": "resetPassword",
                "requestBody": {
                    "$ref": "#/components/requestBodies/ResetPasswordRequest"
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            }
        },
...
    },
    "components": {
        "requestBodies": {
            "LoginRequest": {
                "description": "Login users.",
                "required": true,
                "content": {
                    "application/x-www-form-urlencoded": {
                        "schema": {
                            "required": [
                                "email",
                                "password"
                            ],
                            "properties": {
                                "email": {
                                    "description": "The user email.",
                                    "type": "string",
                                    "format": "email",
                                    "example": "user@email.com"
                                },
                                "password": {
                                    "description": "The user password.",
                                    "type": "string",
                                    "example": "123456"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "RegisterRequest": {
                "description": "Register users.",
                "required": true,
                "content": {
                    "application/x-www-form-urlencoded": {
                        "schema": {
                            "required": [
                                "email",
                                "username",
                                "birthday",
                                "password",
                                "password_confirmation"
                            ],
                            "properties": {
                                "email": {
                                    "description": "The user email.",
                                    "type": "string",
                                    "example": "user@email.com"
                                },
                                "username": {
                                    "description": "The username.",
                                    "type": "string",
                                    "example": "username"
                                },
                                "birthday": {
                                    "description": "The user birthday.",
                                    "type": "string",
                                    "example": "2000-01-01"
                                },
                                "password": {
                                    "description": "The user password.",
                                    "type": "string",
                                    "format": "password",
                                    "example": "123456"
                                },
                                "password_confirmation": {
                                    "description": "The user password confirmation.",
                                    "type": "string",
                                    "format": "password",
                                    "example": "123456"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "RequestResetPasswordRequest": {
                "description": "Reset user password request.",
                "required": true,
                "content": {
                    "application/x-www-form-urlencoded": {
                        "schema": {
                            "required": [
                                "email"
                            ],
                            "properties": {
                                "email": {
                                    "description": "The user email.",
                                    "type": "string",
                                    "format": "email",
                                    "example": "user@email.com"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "ResetPasswordRequest": {
                "description": "Reset user password request.",
                "required": true,
                "content": {
                    "application/x-www-form-urlencoded": {
                        "schema": {
                            "required": [
                                "token",
                                "email",
                                "password",
                                "password_confirmation"
                            ],
                            "properties": {
                                "token": {
                                    "description": "The token.",
                                    "type": "string",
                                    "example": "jhkl31xfdc34g1fds5g1351gdf351dfg"
                                },
                                "email": {
                                    "description": "The user email.",
                                    "type": "string",
                                    "example": "user@email.com"
                                },
                                "password": {
                                    "description": "The user new password.",
                                    "type": "string",
                                    "format": "password",
                                    "example": "123456"
                                },
                                "password_confirmation": {
                                    "description": "The user new password confirmation.",
                                    "type": "string",
                                    "format": "password",
                                    "example": "123456"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}